+++ noodles-bgzf/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## 0.13.0 - 2022-07-05
 
 ### Changed